Move server secrets (passwords, etc) to an extra file, so that devs can easily switch between test and deployment environment.
REMEMBER to never check the actual deployment secrets into Git! (Or to change them
after you accidentally did.)

Note that the particular code used here is not that great. It's better to not just move the credentials, but the whole process of establishing a database connection to a separate module.