Livening
Changes required to get it to work on the server.. Forgot how finiky
older versions of PHP were. And also, with Cloudflare caching the
absolute *(&^ out of things, we actually need to "fight" them to get new
files loaded, especially javascript. Should be able to remove the url
parameters in a day or so, but for now, they are needed.

Had to turn off PrettyPrint JSON as the PHP version failed, said it was
a string, not a flag.

Had to remove parameterized array returns..

Had to correct some paths, forcing with dirname(__FILE__) to get correct
URI.

Had to remove the parseJSON things.. they seemed to work before.. but
are not actually necessary.

Had to add the table "templates", otherwise the admin view won't work..
;-)