More voting stuff
Adding new types of votes & disabling dragging.. it was dragging on
EVERY part of the widget.. not just the top.. unsure how to make it just
drag one div, as it seems to simply drag the whole thing.. ie, if I put
a div in the header of the widget, ONLY THE HEADER WILL DRAG..
irritating. I need a "handle" to grab that lets you drag the whole
widget.. will think on it.