Proposed change to lyricless notification
Noticed the missing lyrics system required user intervention in order to
generate an email to RT, this change would initiate the email
automatically upon a sing user visiting a song page that was lacking
lyrics. With the new widget system in place, this should rarely happen,
that is why this is in a proposed branch.

Code changes lyrics so notification would only ever be issued once, by
inserting an invisible HTML entity into the lyrics.

Should we also initiate a reddit message to the bandit specified as
"lyricist"?